+++ include/drivers/compiler.h
@@ -1,6 +1,6 @@
 #ifndef EXPR_COMPILER_H
 #define EXPR_COMPILER_H
-#include "parser/driver.h"
+#include "drivers/driver.h"
 
 namespace expr {
     class compiler : public driver {
