+++ include/drivers/interpreter.h
@@ -1,7 +1,7 @@
 #ifndef EXPR_INTERPRETER_H
 #define EXPR_INTERPRETER_H
 #include "operations.h"
-#include "parser/driver.h"
+#include "drivers/driver.h"
 
 namespace expr {
     struct interpreter : public driver, arithmetic_operator, boolean_operator, compare_operator {
